starknet: move 'convert' to 'core' directory
