+++ packages/starknet/lib/src/crypto/derivation.dart
@@ -4,7 +4,7 @@ import 'package:bip32/bip32.dart' as bip32;
 import 'package:bip39/bip39.dart' as bip39;
 import 'package:pointycastle/digests/sha256.dart';
 
-import '../convert.dart';
+import '../core/convert.dart';
 import '../core/types/felt.dart';
 import 'model/pedersen_params.dart';
