+++ packages/starknet/lib/src/crypto/keccak.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 import 'dart:typed_data';
 import 'package:pointycastle/digests/keccak.dart';
 
-import '../convert.dart';
+import '../core/convert.dart';
 import '../core/types/index.dart';
 
 /// [mask250] is used to make sure a big int has less than 250 digits.
